:art: zv: refactor serializer pieces into modules
This change is focused on improving the readability of the dbus
and gvariant  serializer code by simply pulling its significant
components out into several smaller files as a prelude to adding
more features and further changes/fixes to this code.

The dbus and gvariant implementations are very similiar in
structure. The same changes are essentially mirrored in the
two implementations.

There are no functional changes here. The API exposed publically
is still the same. These reorganizations are strictly internal to
the dbus and gvariant serializer implementations.