:art: zv: split seq serialization into array+dict
This change splits the DBus and GVariant implementations of
SeqSerializer into an ArraySerializer and DictSerializer. While
these things are pretty similiar - a dict is just an array of
DictEntry in the wire format - the semantics and interface seen
by serde are different since we are talking about individual
elements vs key/value pairs. Splitting this aids readability
and also will facilitate a growing divergence in behavior
between the two as more dict semantics are supported by this core
serializer.

There are no real functionality changes here. This is largely
a refactor of the existing behavior.