:sparkles: zv: support serializing structs with serde
To support serializing structs and nested structs a change was
needed to peek ahead at the array type for a dict (i.e. `a{sv}`)
type as opposed to some other array type (i.e. `ai`). When the
peek sees that this array is an array of dict key/value entries
it now invokes the DictSerializer which properly handles this
case. To this end a `StructSeqSerializer::Dict` was added and
various callpoints were modifie to handle this.

The result is that structs seem to trivially work for both
serialization or deserialization - but optional fields
are less straightforward. These will require more work.